Add simple smoke test
Problem:
We want to ensure, that the pict command is working in the build docker
image.

Solution:
Run the container and grep the output for a test string. When the
string is not found, the build should brake.

Use a wrong test string to provoke an error to ensure, that the build
will break on a real error.