Remove mvnw
Problem:
The maven wrapper is not needed, because we expect that maven is
installed on dev machines. And we know, that maven is installed during
the docker build process. Hence we don't need the additional build time
from the maven wrapper, and we also don't want to check in the
maven-wrapper.jar etc.

Solution:
Remove maven wrapper.