Adjust README.md
Problem:
The README.md needs to clarify, why the web frontend only works on
desktop viewports. In addition it should be explained how to build and
run the application in a docker image locally.

Solution:
Add said adjustments to the documentation. Also add a screenshot to give
in impression of the web frontend without the need to build the project.