Add build time and commit hash info to docker image
Problem:
We can't see only from the build tag which application version is
currently deployed. We want to provide this information in a health
check. Also the Pict version should not be hardcoded in the html.

Solution:
As a first step provide the build time and the commit hash as docker
environment variables. By inspecting the image, we see which version is
currently running.
In a next step we will read these values in the Java code and provide
the information to the frontend.