feat: add CycloneDX Software Bill of Materials to Docker image
Problem:
We want to include a Software Bill of Materials (SBOM) in our Docker
image, using the CycloneDX format, to provide transparency about our
software components.

Solution:
Add the CycloneDX Maven plugin to the pom.xml to generate the SBOM files
during the package phase. Copy the generated SBOM files to the root
directory in the Docker image.

Also add a section to the README.md to inform users about the inclusion
of SBOMs in the Docker image and how to use them.  # Please enter the
commit message for your changes. Lines starting