docs: Add link to website and update Docker run command
Problem:
The README file lacked a link to the project's website, and did not
explain how to use the images published on Docker Hub.

Solution:
Added a link to the website at the top of the README file, and
added a Docker run command which uses the latest image from Docker Hub.